Update README.md
ProgettoBridges is a non-profit, social and cultural promotion association based in Rome. It proposes itself as an intermediary between citizenship, institutions and businesses for the dissemination and development of technologies and cultures related to Blockchain, NFTs and WEB 3.0 through the retrieval of the best national, community and international professionals with whom to build a network in perpetual expansion, in order to improve and speed up its synergies.